move types into different files
