+++ src/types/index.ts
@@ -1,0 +1,9 @@
+import { ManagedIdentityTokenSource } from "../tokensource";
+
+export * from "./activityTypes";
+export * from "./durableClientTypes";
+export * from "./entityTypes";
+export * from "./orchestrationTypes";
+
+// Over time we will likely add more implementations
+export type TokenSource = ManagedIdentityTokenSource;
